refactor: tmp/tmp.py into core/config.py
