+++ output_generators/plaintext.py
@@ -1,7 +1,7 @@
 import os
 from pathlib import Path
 
-import tmp.tmp as tmp
+import core.config as tmp
 
 
 def write_plaintext(microservices, information_flows, external_components):
