+++ output_generators/traceability.py
@@ -2,7 +2,7 @@ import json
 import os
 from pathlib import Path
 
-import tmp.tmp as tmp
+import core.config as tmp
 
 traceability = dict()
 traceability["nodes"] = dict()
