+++ technology_specific_extractors/docker_compose/dcm_parser.py
@@ -6,7 +6,7 @@ import ruamel.yaml
 
 import output_generators.traceability as traceability
 import technology_specific_extractors.environment_variables as env
-import tmp.tmp as tmp
+import core.config as tmp
 
 
 # The following is taken from ruamel.yaml's authro as a workaround for getting line count for str objects
