+++ technology_specific_extractors/html/html_entry.py
@@ -3,7 +3,7 @@ import ast
 import core.file_interaction as fi
 import core.technology_switch as tech_sw
 import output_generators.traceability as traceability
-import tmp.tmp as tmp
+import core.config as tmp
 
 
 def set_information_flows(dfd):
