+++ technology_specific_extractors/implicit_connections/imp_entry.py
@@ -6,7 +6,7 @@ import yaml
 import core.file_interaction as fi
 from output_generators.logger import logger
 import core.technology_switch as tech_sw
-import tmp.tmp as tmp
+import core.config as tmp
 import output_generators.traceability as traceability
