+++ technology_specific_extractors/kafka/kfk_entry.py
@@ -5,7 +5,7 @@ import yaml
 
 import core.file_interaction as fi
 import core.technology_switch as tech_sw
-import tmp.tmp as tmp
+import core.config as tmp
 import output_generators.traceability as traceability
 from output_generators.logger import logger
