+++ technology_specific_extractors/maven/mvn_entry.py
@@ -7,7 +7,7 @@ from output_generators.logger import logger
 import core.parse_files as parse
 import core.technology_switch as tech_sw
 import technology_specific_extractors.docker.dcr_entry as dcr
-import tmp.tmp as tmp
+import core.config as tmp
 import output_generators.traceability as traceability
 
 try:
