+++ technology_specific_extractors/nginx/ngn_entry.py
@@ -4,7 +4,7 @@ import os
 import core.external_components as ext
 import core.file_interaction as fi
 import core.technology_switch as tech_sw
-import tmp.tmp as tmp
+import core.config as tmp
 import output_generators.traceability as traceability
