+++ technology_specific_extractors/spring_config/cnf_entry.py
@@ -4,7 +4,7 @@ import core.file_interaction as fi
 import core.parse_files as parse
 import technology_specific_extractors.environment_variables as env
 import core.technology_switch as tech_sw
-import tmp.tmp as tmp
+import core.config as tmp
 import output_generators.traceability as traceability
