+++ technology_specific_extractors/turbine/trb_entry.py
@@ -2,7 +2,7 @@ import os
 
 import core.file_interaction as fi
 import core.technology_switch as tech_sw
-import tmp.tmp as tmp
+import core.config as tmp
 import output_generators.traceability as traceability
