+++ technology_specific_extractors/zuul/zul_entry.py
@@ -1,7 +1,7 @@
 import core.external_components as ext
 import core.file_interaction as fi
 import core.technology_switch as tech_sw
-import tmp.tmp as tmp
+import core.config as tmp
 import output_generators.traceability as traceability
